Non working code
In the debugger the value of MINUTEFLAG changes, however in the while
loop the value of MINUTEFLAG is never not 1. When running the debugger
the value of MINUTE flag goes to 1 but then it does not call the
clockUpdate function. Not sure why this is happening. Also in the debug
the stack pointer gets lost at certain points and the onboard LEDs do
not light up.